Added event system, fixed cleanup timing
Cleanup is now called after the game loop has actually finished and not
when game_stop is called. This fixes cleanup hooks getting called
potentially before the game is actually finished, leading to possible
segfaults.

Added an event system, with the ability to add listeners that can handle
multiple event types, using bit flags. To check what even type it is,
use event.type, and then it can be casted to the appropiate event struct.